+++ lib/sidekiq/worker.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 require 'sidekiq/client'
-require 'sidekiq/core_ext'
 
 module Sidekiq
 
@@ -28,13 +27,42 @@ module Sidekiq
       raise ArgumentError, "You cannot include Sidekiq::Worker in an ActiveJob: #{base.name}" if base.ancestors.any? {|c| c.name == 'ActiveJob::Base' }
 
       base.extend(ClassMethods)
-      base.class_attribute :sidekiq_options_hash
-      base.class_attribute :sidekiq_retry_in_block
-      base.class_attribute :sidekiq_retries_exhausted_block
+      base.sidekiq_class_attribute :sidekiq_options_hash
+      base.sidekiq_class_attribute :sidekiq_retry_in_block
+      base.sidekiq_class_attribute :sidekiq_retries_exhausted_block
     end
 
     def logger
       Sidekiq.logger
+    end
+
+    # This helper class encapsulates the set options for `set`, e.g.
+    #
+    #     SomeWorker.set(queue: 'foo').perform_async(....)
+    #
+    class Setter
+      def initialize(klass, opts)
+        @klass = klass
+        @opts = opts
+      end
+
+      def perform_async(*args)
+        @klass.client_push(@opts.merge('args'.freeze => args, 'class'.freeze => @klass))
+      end
+
+      # +interval+ must be a timestamp, numeric or something that acts
+      #   numeric (like an activesupport time interval).
+      def perform_in(interval, *args)
+        int = interval.to_f
+        now = Time.now.to_f
+        ts = (int < 1_000_000_000 ? now + int : int)
+
+        payload = @opts.merge('class'.freeze => @klass, 'args'.freeze => args, 'at'.freeze => ts)
+        # Optimization to enqueue something now that is scheduled to go out now or in the past
+        payload.delete('at'.freeze) if ts <= now
+        @klass.client_push(payload)
+      end
+      alias_method :perform_at, :perform_in
     end
 
     module ClassMethods
@@ -52,12 +80,11 @@ module Sidekiq
       end
 
       def set(options)
-        Thread.current[:sidekiq_worker_set] = options
-        self
+        Setter.new(self, options)
       end
 
       def perform_async(*args)
-        client_push('class' => self, 'args' => args)
+        client_push('class'.freeze => self, 'args'.freeze => args)
       end
 
       # +interval+ must be a timestamp, numeric or something that acts
@@ -67,7 +94,7 @@ module Sidekiq
         now = Time.now.to_f
         ts = (int < 1_000_000_000 ? now + int : int)
 
-        item = { 'class' => self, 'args' => args, 'at' => ts }
+        item = { 'class'.freeze => self, 'args'.freeze => args, 'at'.freeze => ts }
 
         # Optimization to enqueue something now that is scheduled to go out now or in the past
         item.delete('at'.freeze) if ts <= now
@@ -90,7 +117,8 @@ module Sidekiq
       # In practice, any option is allowed.  This is the main mechanism to configure the
       # options for a specific job.
       def sidekiq_options(opts={})
-        self.sidekiq_options_hash = get_sidekiq_options.merge(opts.stringify_keys)
+        # stringify
+        self.sidekiq_options_hash = get_sidekiq_options.merge(Hash[opts.map{|k, v| [k.to_s, v]}])
       end
 
       def sidekiq_retry_in(&block)
@@ -106,14 +134,69 @@ module Sidekiq
       end
 
       def client_push(item) # :nodoc:
-        pool = Thread.current[:sidekiq_via_pool] || get_sidekiq_options['pool'] || Sidekiq.redis_pool
-        hash = if Thread.current[:sidekiq_worker_set]
-          x, Thread.current[:sidekiq_worker_set] = Thread.current[:sidekiq_worker_set], nil
-          x.stringify_keys.merge(item.stringify_keys)
-        else
-          item.stringify_keys
+        pool = Thread.current[:sidekiq_via_pool] || get_sidekiq_options['pool'.freeze] || Sidekiq.redis_pool
+        # stringify
+        item.keys.each do |key|
+          item[key.to_s] = item.delete(key)
         end
-        Sidekiq::Client.new(pool).push(hash)
+
+        Sidekiq::Client.new(pool).push(item)
+      end
+
+      def sidekiq_class_attribute(*attrs)
+        instance_reader = true
+        instance_writer = true
+
+        attrs.each do |name|
+          singleton_class.instance_eval do
+            undef_method(name) if method_defined?(name) || private_method_defined?(name)
+          end
+          define_singleton_method(name) { nil }
+
+          ivar = "@#{name}"
+
+          singleton_class.instance_eval do
+            m = "#{name}="
+            undef_method(m) if method_defined?(m) || private_method_defined?(m)
+          end
+          define_singleton_method("#{name}=") do |val|
+            singleton_class.class_eval do
+              undef_method(name) if method_defined?(name) || private_method_defined?(name)
+              define_method(name) { val }
+            end
+
+            if singleton_class?
+              class_eval do
+                undef_method(name) if method_defined?(name) || private_method_defined?(name)
+                define_method(name) do
+                  if instance_variable_defined? ivar
+                    instance_variable_get ivar
+                  else
+                    singleton_class.send name
+                  end
+                end
+              end
+            end
+            val
+          end
+
+          if instance_reader
+            undef_method(name) if method_defined?(name) || private_method_defined?(name)
+            define_method(name) do
+              if instance_variable_defined?(ivar)
+                instance_variable_get ivar
+              else
+                self.class.public_send name
+              end
+            end
+          end
+
+          if instance_writer
+            m = "#{name}="
+            undef_method(m) if method_defined?(m) || private_method_defined?(m)
+            attr_writer name
+          end
+        end
       end
 
     end