+++ test/test_client.rb
@@ -23,6 +23,10 @@ class TestClient < Sidekiq::Test
       assert_raises ArgumentError do
         Sidekiq::Client.push('queue' => 'foo', 'class' => MyWorker, 'args' => 1)
       end
+
+      assert_raises ArgumentError do
+        Sidekiq::Client.push('queue' => 'foo', 'class' => MyWorker, 'args' => [1], 'at' => Time.now)
+      end
     end
   end
 
@@ -234,6 +238,19 @@ class TestClient < Sidekiq::Test
 
     def setup
       Sidekiq.redis {|c| c.flushdb }
+    end
+
+    it 'can be memoized' do
+      q = Sidekiq::Queue.new('bar')
+      assert_equal 0, q.size
+      set = SetWorker.set(queue: :bar, foo: 'qaaz')
+      set.perform_async(1)
+      set.perform_async(1)
+      set.perform_async(1)
+      set.perform_async(1)
+      assert_equal 4, q.size
+      assert_equal 4, q.map{|j| j['jid'] }.uniq.size
+      set.perform_in(10, 1)
     end
 
     it 'allows option overrides' do