+++ test/test_middleware.rb
@@ -82,6 +82,7 @@ class TestMiddleware < Sidekiq::Test
       boss = Minitest::Mock.new
       boss.expect(:options, {:queues => ['default'] }, [])
       boss.expect(:options, {:queues => ['default'] }, [])
+      boss.expect(:options, {:queues => ['default'] }, [])
       processor = Sidekiq::Processor.new(boss)
       boss.expect(:processor_done, nil, [processor])
       processor.process(Sidekiq::BasicFetch::UnitOfWork.new('queue:default', msg))
@@ -90,8 +91,8 @@ class TestMiddleware < Sidekiq::Test
 
     it 'correctly replaces middleware when using middleware with options in the initializer' do
       chain = Sidekiq::Middleware::Chain.new
-      chain.add Sidekiq::Middleware::Server::RetryJobs
-      chain.add Sidekiq::Middleware::Server::RetryJobs, {:max_retries => 5}
+      chain.add NonYieldingMiddleware
+      chain.add NonYieldingMiddleware, {:foo => 5}
       assert_equal 1, chain.count
     end