+++ test/test_retry_exhausted.rb
@@ -1,13 +1,13 @@
 # encoding: utf-8
 require_relative 'helper'
-require 'sidekiq/middleware/server/retry_jobs'
+require 'sidekiq/job_retry'
 
 class TestRetryExhausted < Sidekiq::Test
   describe 'sidekiq_retries_exhausted' do
     class NewWorker
       include Sidekiq::Worker
 
-      class_attribute :exhausted_called, :exhausted_job, :exhausted_exception
+      sidekiq_class_attribute :exhausted_called, :exhausted_job, :exhausted_exception
 
       sidekiq_retries_exhausted do |job, e|
         self.exhausted_called = true
@@ -19,7 +19,7 @@ class TestRetryExhausted < Sidekiq::Test
     class OldWorker
       include Sidekiq::Worker
 
-      class_attribute :exhausted_called, :exhausted_job, :exhausted_exception
+      sidekiq_class_attribute :exhausted_called, :exhausted_job, :exhausted_exception
 
       sidekiq_retries_exhausted do |job|
         self.exhausted_called = true
@@ -52,7 +52,7 @@ class TestRetryExhausted < Sidekiq::Test
     end
 
     def handler(options={})
-      @handler ||= Sidekiq::Middleware::Server::RetryJobs.new(options)
+      @handler ||= Sidekiq::JobRetry.new(options)
     end
 
     def job(options={})
@@ -60,62 +60,64 @@ class TestRetryExhausted < Sidekiq::Test
     end
 
     it 'does not run exhausted block when job successful on first run' do
-      handler.call(new_worker, job('retry' => 2), 'default') do
+      handler.local(new_worker, job('retry' => 2), 'default') do
         # successful
       end
 
-      refute NewWorker.exhausted_called?
+      refute NewWorker.exhausted_called
     end
 
     it 'does not run exhausted block when job successful on last retry' do
-      handler.call(new_worker, job('retry_count' => 0, 'retry' => 1), 'default') do
+      handler.local(new_worker, job('retry_count' => 0, 'retry' => 1), 'default') do
         # successful
       end
 
-      refute NewWorker.exhausted_called?
+      refute NewWorker.exhausted_called
     end
 
     it 'does not run exhausted block when retries not exhausted yet' do
       assert_raises RuntimeError do
-        handler.call(new_worker, job('retry' => 1), 'default') do
+        handler.local(new_worker, job('retry' => 1), 'default') do
           raise 'kerblammo!'
         end
       end
 
-      refute NewWorker.exhausted_called?
+      refute NewWorker.exhausted_called
     end
 
     it 'runs exhausted block when retries exhausted' do
       assert_raises RuntimeError do
-        handler.call(new_worker, job('retry_count' => 0, 'retry' => 1), 'default') do
+        handler.local(new_worker, job('retry_count' => 0, 'retry' => 1), 'default') do
           raise 'kerblammo!'
         end
       end
 
-      assert NewWorker.exhausted_called?
+      assert NewWorker.exhausted_called
     end
 
 
     it 'passes job and exception to retries exhausted block' do
       raised_error = assert_raises RuntimeError do
-        handler.call(new_worker, job('retry_count' => 0, 'retry' => 1), 'default') do
+        handler.local(new_worker, job('retry_count' => 0, 'retry' => 1), 'default') do
           raise 'kerblammo!'
         end
       end
+      raised_error = raised_error.cause
 
-      assert new_worker.exhausted_called?
+      assert new_worker.exhausted_called
       assert_equal raised_error.message, new_worker.exhausted_job['error_message']
       assert_equal raised_error, new_worker.exhausted_exception
     end
 
     it 'passes job to retries exhausted block' do
       raised_error = assert_raises RuntimeError do
-        handler.call(old_worker, job('retry_count' => 0, 'retry' => 1), 'default') do
+        handler.local(old_worker, job('retry_count' => 0, 'retry' => 1), 'default') do
           raise 'kerblammo!'
         end
       end
+      raised_error = raised_error.cause
 
-      assert old_worker.exhausted_called?
+      assert old_worker.exhausted_called
       assert_equal raised_error.message, old_worker.exhausted_job['error_message']
       assert_nil new_worker.exhausted_exception
     end
@@ -134,10 +136,11 @@ class TestRetryExhausted < Sidekiq::Test
         end
         f = Foobar.new
         raised_error = assert_raises RuntimeError do
-          handler.call(f, job('retry_count' => 0, 'retry' => 1), 'default') do
+          handler.local(f, job('retry_count' => 0, 'retry' => 1), 'default') do
             raise 'kerblammo!'
           end
         end
+        raised_error = raised_error.cause
 
         assert exhausted_job
         assert_equal raised_error, exhausted_exception